Закончил с первой анкетой. Начал вторую 'documents'

+++ release/views.py
@@ -1,12 +1,10 @@
-from django import forms
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.auth.models import User
-from django.http import request
-from django.http.response import HttpResponse, HttpResponseRedirect
-from django.shortcuts import render, resolve_url
+from django.forms import formset_factory
+from django.http.response import HttpResponseRedirect
+from django.shortcuts import render
 from django.urls.base import reverse_lazy
 from django.views.generic.edit import FormView
-from django.forms import formset_factory
 
 from .forms import *
 from .models import *
@@ -18,13 +16,13 @@ class MainInfoView(LoginRequiredMixin, FormView):
 
     def get(self, request, *args, **kwargs):
         user = User.objects.get(username = request.user)
-        try:
-            main_info = MainInfo.objects.filter(user_id = user.id).values()
+        main_info = MainInfo.objects.filter(user_id = user.id).values()
+        if len(main_info) != 0:
             for item in main_info:
-                main = item
-            main['user'] = user
-            form = self.form_class(initial = main)
-        except:
+                data = item
+            data['user'] = user
+            form = self.form_class(initial = data)
+        else:
             form = self.form_class(initial = {'user': user})
         return render(request, self.template_name, {'form': form})
 
@@ -32,12 +30,10 @@ class MainInfoView(LoginRequiredMixin, FormView):
         form = self.form_class(data = request.POST, files = request.FILES)
         user = User.objects.get(username = request.user)
         if form.is_valid():
-            try:
-                main_info = MainInfo.objects.filter(user_id = user.id)
+            main_info = MainInfo.objects.filter(user_id = user.id)
+            if len(main_info) != 0:
                 for item in main_info:
                     item.delete()
-            except:
-                pass
             form.save()
             if form.cleaned_data['content_type'] == 'SINGLE' or form.cleaned_data['content_type'] == 'ALBUM':
                 self.next = 'audio'
@@ -76,6 +72,7 @@ class AudioView(LoginRequiredMixin, FormView):
         user = User.objects.get(username = request.user)
         audio_formset = formset_factory(self.form_class)
         formset = audio_formset(data = self.request.POST, files = self.request.FILES)
+        add_video = self.request.POST.get('add_video')
         if formset.is_valid():
             try:
                 audio = Audio.objects.filter(user_id = user.id)
@@ -85,10 +82,44 @@ class AudioView(LoginRequiredMixin, FormView):
                 pass
             for form in formset:
                 form.save()
-            return HttpResponseRedirect(reverse_lazy('success'))
+            if add_video == 'NO':
+                return HttpResponseRedirect(reverse_lazy('success'))
+            else:
+                return HttpResponseRedirect(reverse_lazy('video'))
         else:
             return render(request, self.template_name, {'button': self.fields_for_button, 'formset': formset})
 
+
+class VideoView(LoginRequiredMixin, FormView):
+    template_name = 'release/video.html'
+    form_class = VideoForm
+
+    def get(self, request, *args, **kwargs):
+        user = User.objects.get(username = request.user)
+        video = Video.objects.filter(user_id = user.id).values()
+        if len(video) != 0:
+            for item in video:
+                data = item
+            data['user'] = user
+            form = self.form_class(initial = data)
+        else:
+            form = self.form_class(initial = {'user': user})
+        return render(request, self.template_name, {'form': form})
+
+    def post(self, request, *args, **kwargs):
+        form = self.form_class(data = request.POST, files = request.FILES)
+        user = User.objects.get(username = request.user)
+        if form.is_valid():
+            video = Video.objects.filter(user_id = user.id)
+            if len(video) != 0:
+                for item in video:
+                    item.delete()
+            form.save()
+            return HttpResponseRedirect(reverse_lazy('success'))
+        else:
+            return render(request, self.template_name, {'form': form})
+
+            
 
 def success_page(request):
     return render(request, 'release/success.html')
